requirements.txt
absl-py==2.2.2
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
audiomentations==0.41.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
keyboard==0.13.5
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.9
narwhals==1.38.0
numba==0.61.2
numpy==2.1.3
numpy-minmax==0.4.0
numpy-rms==0.5.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==5.29.4
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-stretch==0.3.1
python-vlc==3.0.21203
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.3.1
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
streamlit==1.45.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2